refactor state somewhat
