+++ src/pages/plan/parameters/DurationSelector.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useState } from "react";
-import { durationDaysState } from "../../../store/dates";
 import Button from "react-bootstrap/esm/Button";
 import InputGroup from "react-bootstrap/InputGroup";
 import "./DurationSelector.css";
+import { durationDaysState } from "../../../store/displayDateRange";
 
 export const DurationSelector = () => {
   const [queryRange, setQueryRange] = useState(durationDaysState.peek());
